# src/legal-document-simplifier/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Legal Document Simplifier API",
    "description": "AI-powered legal document analysis with risk scoring and RAG-based Q&A",
    "version": "v1"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "description": "Health check endpoint for Cloud Run",
        "operationId": "health_check_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "root"
        ],
        "summary": "Root",
        "description": "API root endpoint",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/upload/": {
      "post": {
        "tags": [
          "upload"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_upload__post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/analyze/": {
      "post": {
        "tags": [
          "analyze"
        ],
        "summary": "Analyze Document",
        "operationId": "analyze_document_analyze__post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/chat/": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_chat__post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}